% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountGenome-package.R
\docType{package}
\name{SeqCountGenome-package}
\alias{SeqCountGenome-package}
\title{SeqCountGenome - Counting sequences occurrences in subsets of the genome}
\description{
The \code{SeqCountGenome} package implements a function for counting
the occurrences of a set of search sequences in specific regions of the
genome. The occurrences are found through exact matching and reported in
a count matrix.
}
\details{
\tabular{ll}{
Package: \tab SeqCountGenome\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2021-07-13\cr
License: \tab GPL (>=2)\cr
}
}
\references{
\url{https://github.com/dacolombo/SeqCountGenome}\cr
}
\author{
Daniele Colombo\cr
Politecnico di Milano\cr
Maintainer: Daniele Colombo\cr
E-Mail: \href{mailto:daniele20.colombo@mail.polimi.it}{daniele20.colombo@mail.polimi.it}
}
