% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountGenome.R
\name{SeqCountGenome}
\alias{SeqCountGenome}
\title{Count sequences occurrences in subsets of the genome}
\usage{
SeqCountGenome(genome,
regions,
patterns,
overall.count = FALSE,
remove.0.matches = FALSE)
}
\arguments{
\item{genome}{A BSgenome object corresponding to the reference genome}

\item{regions}{A GRanges object containing the genomic regions of interest}

\item{patterns}{A character vector containing the search sequences}

\item{overall.count}{A logical, if TRUE the sum of matches for each search
sequence will be computed}

\item{remove.0.matches}{A logical, if TRUE the regions with no matches will
be removed from the output matrix}
}
\value{
A matrix containing the exact matches counts, with search sequences
as columns and genomic regions as rows
}
\description{
This function find exact matches of search sequences in subsets of the
genome. The reference genome, the regions of interest and a character vector
containing the search sequences must be passed. Optionally, regions for
which no match is found can be dropped and the total count of matches can be
computed for each search sequence. A count matrix is returned,
with search sequences as columns and genomic regions as rows.
}
\examples{
# Reference genome: hg38
Hsapiens <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
# Genomic regions: promoters
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
promoters <- promoters(txdb, upstream=2000, downstream=200)
promoters <- promoters[grep("_",seqnames(promoters),invert=TRUE)]
# Search patterns: transcription factor consensus
transcription_consensus <- c("GTCAAGGTCA","CACGTG","GGGCGG")

SeqCountGenome(Hsapiens,promoters,transcription_consensus)


}
\author{
Daniele Colombo\cr
Politecnico di Milano\cr
Maintainer: Daniele Colombo\cr
E-Mail: \href{mailto:daniele20.colombo@mail.polimi.it}{daniele20.colombo@mail.polimi.it}
}
